--- .github/FUNDING.yml
@@ -1,3 +1,0 @@
-github: ['schorschii']
-liberapay: schorschii
-custom: ['https://www.paypal.me/schorschii', 'https://play.google.com/store/apps/details?id=systems.sieber.remotespotlight']
