+++ .gitignore
@@ -1,8 +1,4 @@
-**/nbproject/private/
-**/nbproject/Makefile-*.mk
-**/nbproject/Package-*.bash
-build/
-nbbuild/
-dist/
-nbdist/
-.nb-gradle/
+
+.vs/
+
+CRemotePointerServer/out/
