+++ CRemotePointerServer/listener.h
@@ -1,0 +1,4 @@
+#pragma once
+
+void create_listener(int port, const char* connection_code);
+
