+++ README.md
@@ -1,3 +1,8 @@
+# CRemotePointer-Server
+
+C port of the Java RemotePointer-Server.
+
+
 # RemotePointer-Server
 Server application for the [RemotePointer Android App](https://github.com/schorschii/RemotePointer-Android). With RemotePointer you can use your smartphone to control your computer's mouse and keyboard.
