--- src/RemotePointerServer/Bundle.properties
@@ -1,21 +1,0 @@
-# To change this license header, choose License Headers in Project Properties.
-# To change this template file, choose Tools | Templates
-# and open the template in the editor.
-
-SetupWindow.jLabelCode.text=0000
-SetupWindow.title=RemotePointer Server
-SetupWindow.jLabelHostname.AccessibleContext.accessibleName=Connection Info
-SetupWindow.jLabelHostname.text=Hostname
-SetupWindow.jButtonExit.text=Exit
-AboutWindow.jTextAreaAbout.text=Please download the Android-App RemotePointer from Google Play to control your computer with this software.\n\nFree Software!! Pull requests welcome: https://github.com/schorschii/RemotePointer-Server\n\nCommand Line Parameters:\n --code=<code>  -  use <code> as connection code\n --hostname=<name>  -  use <name> as hostname\n --listenPort=<port>  -  use <port> as server port (default 4444)\n --broadcastPort=<port>  -  use <port> for broadcast messages (default 4445)\n --hidden  -  start without showing the main window (for autostart)\n\n\u00a9 Georg Sieber 2018-2019\nhttps://georg-sieber.de\nit@georg-sieber.de\n\nUsing Material Icons\nhttps://material.io/tools/icons/
-AboutWindow.title=About
-AboutWindow.jLabelPointerSpeed.text=Pointer Speed:
-AboutWindow.jLabelVersion.text=v2.3
-AboutWindow.jLabelMouseSpeed.text=Mouse Speed:
-AboutWindow.jButtonPlayStore.text=Visit Play Store
-AboutWindow.jButtonWebsite.text=Visit Developer's Website
-AboutWindow.jButtonClose.text=Close
-AboutWindow.jLabelPointerSpeed1.text=Pointer Style:
-AboutWindow.jButtonClose1.text=Exit And Generate New Connection Code
-TrayMenu.Exit=Exit
-SetupWindow.jButtonInfo.text=Settings / Help / About
