--- src/RemotePointerServer/Bundle_de_DE.properties
@@ -1,17 +1,0 @@
-
-SetupWindow.jLabelCode.text=0000
-SetupWindow.title=RemotePointer Server
-SetupWindow.jLabelHostname.AccessibleContext.accessibleName=Verbindungsinformationen
-SetupWindow.jLabelHostname.text=Rechnername
-SetupWindow.jButtonExit.text=Beenden
-AboutWindow.jTextAreaAbout.text=Bitte laden Sie sich die Android-App RemotePointer aus dem Google Play Store herunter, um Ihren Computer mit dieser Software zu steuern.\n\nFreie Software!! Pull requests sind gerne gesehen: https://github.com/schorschii/RemotePointer-Server\n\nKommandozeilenparameter:\n --code=<code>  -  verwende <code> als Verbindungscode\n --hostname=<name>  -  verwende <name> als Hostname\n --listenPort=<port>  -  verwende <port> als Serverport (Standard 4444)\n --broadcastPort=<port>  -  verwende <port> f\u00fcr Broadcast-Nachrichten (Standard 4445)\n --hidden  -  starte, ohne das Hauptfenster anzuzeigen (f\u00fcr Autostart)\n\n\u00a9 Georg Sieber 2018-2019\nhttps://georg-sieber.de\nit@georg-sieber.de\n\nVerwendet die Material Icons\nhttps://material.io/tools/icons/
-AboutWindow.title=Informationen
-AboutWindow.jLabelPointerSpeed.text=Laserpointer-Geschwindigkeit:
-AboutWindow.jLabelMouseSpeed.text=Mauszeiger-Geschwindigkeit:
-AboutWindow.jButtonPlayStore.text=Play Store \u00f6ffnen
-AboutWindow.jButtonWebsite.text=Entwicklerwebsite \u00f6ffnen
-AboutWindow.jButtonClose.text=Schlie\u00dfen
-AboutWindow.jLabelPointerSpeed1.text=Laserpointer-Stil:
-AboutWindow.jButtonClose1.text=Beenden und neuen Verbindungscode generieren
-TrayMenu.Exit=Beenden
-SetupWindow.jButtonInfo.text=Einstellungen / Hilfe / Info
