--- src/RemotePointerServer/SetupWindow.java
@@ -1,232 +1,0 @@
-/*
- * To change this license header, choose License Headers in Project Properties.
- * To change this template file, choose Tools | Templates
- * and open the template in the editor.
- */
-package RemotePointerServer;
-
-import java.io.IOException;
-import java.util.logging.Level;
-import java.util.logging.Logger;
-import javax.imageio.ImageIO;
-import javax.swing.ImageIcon;
-import javax.swing.JFrame;
-
-/**
- *
- * @author georg
- */
-public class SetupWindow extends javax.swing.JFrame {
-
-    private RuntimeSettings rs;
-
-    /**
-     * Creates new form SetupWindow
-     */
-    public SetupWindow() {
-        baseInit();
-    }
-    public SetupWindow(RuntimeSettings rs) {
-        this.rs = rs;
-        baseInit();
-    }
-    private void baseInit() {
-        try {
-            initComponents();
-            this.setDefaultCloseOperation(JFrame.DO_NOTHING_ON_CLOSE);
-            setIconImage(ImageIO.read(SetupWindow.class.getResource("res/ic_launcher_round.png")));
-            if(getBackground().getRed() > 200) jLabelHostnameImg.setIcon(new ImageIcon(SetupWindow.class.getResource("res/b_computer.png")));
-            if(getBackground().getRed() > 200) jLabelCodeImg.setIcon(new ImageIcon(SetupWindow.class.getResource("res/b_key.png")));
-            this.addWindowListener(new java.awt.event.WindowAdapter() {
-                @Override
-                public void windowClosing(java.awt.event.WindowEvent windowEvent) {
-                    setVisible(false);
-                }
-            });
-        } catch (IOException ex) {
-            Logger.getLogger(SetupWindow.class.getName()).log(Level.SEVERE, null, ex);
-        }
-    }
-    
-    public void setText(String hostname, String authCode) {
-        jLabelHostname.setText(hostname);
-        jLabelCode.setText(authCode);
-    }
-
-    /**
-     * This method is called from within the constructor to initialize the form.
-     * WARNING: Do NOT modify this code. The content of this method is always
-     * regenerated by the Form Editor.
-     */
-    @SuppressWarnings("unchecked")
-    // <editor-fold defaultstate="collapsed" desc="Generated Code">//GEN-BEGIN:initComponents
-    private void initComponents() {
-
-        jPanelLabelContainer = new javax.swing.JPanel();
-        jLabelCode = new javax.swing.JLabel();
-        jLabelHostname = new javax.swing.JLabel();
-        jLabelHostnameImg = new javax.swing.JLabel();
-        jLabelCodeImg = new javax.swing.JLabel();
-        jButtonExit = new javax.swing.JButton();
-        jButtonInfo = new javax.swing.JButton();
-
-        setDefaultCloseOperation(javax.swing.WindowConstants.EXIT_ON_CLOSE);
-        java.util.ResourceBundle bundle = java.util.ResourceBundle.getBundle("RemotePointerServer/Bundle"); // NOI18N
-        setTitle(bundle.getString("SetupWindow.title")); // NOI18N
-        setIconImages(null);
-        setName("RemotePointer Server"); // NOI18N
-        setResizable(false);
-
-        jPanelLabelContainer.setPreferredSize(new java.awt.Dimension(250, 250));
-
-        jLabelCode.setFont(new java.awt.Font("Noto Sans", 1, 18)); // NOI18N
-        jLabelCode.setText(bundle.getString("SetupWindow.jLabelCode.text")); // NOI18N
-        jLabelCode.setVerticalTextPosition(javax.swing.SwingConstants.TOP);
-
-        jLabelHostname.setFont(new java.awt.Font("Noto Sans", 1, 18)); // NOI18N
-        jLabelHostname.setText(bundle.getString("SetupWindow.jLabelHostname.text")); // NOI18N
-        jLabelHostname.setVerticalTextPosition(javax.swing.SwingConstants.TOP);
-
-        jLabelHostnameImg.setFont(new java.awt.Font("Noto Sans", 1, 18)); // NOI18N
-        jLabelHostnameImg.setIcon(new javax.swing.ImageIcon(getClass().getResource("/RemotePointerServer/res/computer.png"))); // NOI18N
-        jLabelHostnameImg.setVerticalAlignment(javax.swing.SwingConstants.TOP);
-        jLabelHostnameImg.setVerticalTextPosition(javax.swing.SwingConstants.TOP);
-
-        jLabelCodeImg.setFont(new java.awt.Font("Noto Sans", 1, 18)); // NOI18N
-        jLabelCodeImg.setIcon(new javax.swing.ImageIcon(getClass().getResource("/RemotePointerServer/res/key.png"))); // NOI18N
-        jLabelCodeImg.setVerticalAlignment(javax.swing.SwingConstants.TOP);
-        jLabelCodeImg.setVerticalTextPosition(javax.swing.SwingConstants.TOP);
-
-        javax.swing.GroupLayout jPanelLabelContainerLayout = new javax.swing.GroupLayout(jPanelLabelContainer);
-        jPanelLabelContainer.setLayout(jPanelLabelContainerLayout);
-        jPanelLabelContainerLayout.setHorizontalGroup(
-            jPanelLabelContainerLayout.createParallelGroup(javax.swing.GroupLayout.Alignment.LEADING)
-            .addGroup(jPanelLabelContainerLayout.createSequentialGroup()
-                .addContainerGap()
-                .addGroup(jPanelLabelContainerLayout.createParallelGroup(javax.swing.GroupLayout.Alignment.LEADING)
-                    .addGroup(jPanelLabelContainerLayout.createSequentialGroup()
-                        .addComponent(jLabelHostnameImg)
-                        .addPreferredGap(javax.swing.LayoutStyle.ComponentPlacement.UNRELATED)
-                        .addComponent(jLabelHostname, javax.swing.GroupLayout.DEFAULT_SIZE, javax.swing.GroupLayout.DEFAULT_SIZE, Short.MAX_VALUE))
-                    .addGroup(jPanelLabelContainerLayout.createSequentialGroup()
-                        .addComponent(jLabelCodeImg)
-                        .addPreferredGap(javax.swing.LayoutStyle.ComponentPlacement.UNRELATED)
-                        .addComponent(jLabelCode, javax.swing.GroupLayout.DEFAULT_SIZE, javax.swing.GroupLayout.DEFAULT_SIZE, Short.MAX_VALUE)))
-                .addGap(261, 261, 261))
-        );
-        jPanelLabelContainerLayout.setVerticalGroup(
-            jPanelLabelContainerLayout.createParallelGroup(javax.swing.GroupLayout.Alignment.LEADING)
-            .addGroup(javax.swing.GroupLayout.Alignment.TRAILING, jPanelLabelContainerLayout.createSequentialGroup()
-                .addContainerGap()
-                .addGroup(jPanelLabelContainerLayout.createParallelGroup(javax.swing.GroupLayout.Alignment.LEADING)
-                    .addComponent(jLabelHostname, javax.swing.GroupLayout.PREFERRED_SIZE, 36, javax.swing.GroupLayout.PREFERRED_SIZE)
-                    .addComponent(jLabelHostnameImg))
-                .addPreferredGap(javax.swing.LayoutStyle.ComponentPlacement.RELATED)
-                .addGroup(jPanelLabelContainerLayout.createParallelGroup(javax.swing.GroupLayout.Alignment.TRAILING)
-                    .addComponent(jLabelCodeImg)
-                    .addComponent(jLabelCode, javax.swing.GroupLayout.PREFERRED_SIZE, 36, javax.swing.GroupLayout.PREFERRED_SIZE))
-                .addContainerGap(javax.swing.GroupLayout.DEFAULT_SIZE, Short.MAX_VALUE))
-        );
-
-        jLabelHostname.getAccessibleContext().setAccessibleName(bundle.getString("SetupWindow.jLabelHostname.AccessibleContext.accessibleName")); // NOI18N
-
-        jButtonExit.setText(bundle.getString("SetupWindow.jButtonExit.text")); // NOI18N
-        jButtonExit.setPreferredSize(new java.awt.Dimension(0, 0));
-        jButtonExit.addActionListener(new java.awt.event.ActionListener() {
-            public void actionPerformed(java.awt.event.ActionEvent evt) {
-                jButtonExitActionPerformed(evt);
-            }
-        });
-
-        jButtonInfo.setText(bundle.getString("SetupWindow.jButtonInfo.text")); // NOI18N
-        jButtonInfo.setPreferredSize(new java.awt.Dimension(0, 0));
-        jButtonInfo.addActionListener(new java.awt.event.ActionListener() {
-            public void actionPerformed(java.awt.event.ActionEvent evt) {
-                jButtonInfoActionPerformed(evt);
-            }
-        });
-
-        javax.swing.GroupLayout layout = new javax.swing.GroupLayout(getContentPane());
-        getContentPane().setLayout(layout);
-        layout.setHorizontalGroup(
-            layout.createParallelGroup(javax.swing.GroupLayout.Alignment.LEADING)
-            .addGroup(layout.createSequentialGroup()
-                .addContainerGap()
-                .addComponent(jButtonInfo, javax.swing.GroupLayout.PREFERRED_SIZE, 216, javax.swing.GroupLayout.PREFERRED_SIZE)
-                .addPreferredGap(javax.swing.LayoutStyle.ComponentPlacement.RELATED)
-                .addComponent(jButtonExit, javax.swing.GroupLayout.PREFERRED_SIZE, 121, javax.swing.GroupLayout.PREFERRED_SIZE)
-                .addContainerGap())
-            .addGroup(layout.createSequentialGroup()
-                .addComponent(jPanelLabelContainer, javax.swing.GroupLayout.PREFERRED_SIZE, 367, javax.swing.GroupLayout.PREFERRED_SIZE)
-                .addGap(0, 0, Short.MAX_VALUE))
-        );
-        layout.setVerticalGroup(
-            layout.createParallelGroup(javax.swing.GroupLayout.Alignment.LEADING)
-            .addGroup(javax.swing.GroupLayout.Alignment.TRAILING, layout.createSequentialGroup()
-                .addComponent(jPanelLabelContainer, javax.swing.GroupLayout.PREFERRED_SIZE, 102, javax.swing.GroupLayout.PREFERRED_SIZE)
-                .addPreferredGap(javax.swing.LayoutStyle.ComponentPlacement.RELATED, 16, Short.MAX_VALUE)
-                .addGroup(layout.createParallelGroup(javax.swing.GroupLayout.Alignment.BASELINE)
-                    .addComponent(jButtonExit, javax.swing.GroupLayout.PREFERRED_SIZE, 45, javax.swing.GroupLayout.PREFERRED_SIZE)
-                    .addComponent(jButtonInfo, javax.swing.GroupLayout.PREFERRED_SIZE, 45, javax.swing.GroupLayout.PREFERRED_SIZE))
-                .addContainerGap())
-        );
-
-        pack();
-    }// </editor-fold>//GEN-END:initComponents
-
-    private void jButtonExitActionPerformed(java.awt.event.ActionEvent evt) {//GEN-FIRST:event_jButtonExitActionPerformed
-        System.exit(0);
-    }//GEN-LAST:event_jButtonExitActionPerformed
-
-    private void jButtonInfoActionPerformed(java.awt.event.ActionEvent evt) {//GEN-FIRST:event_jButtonInfoActionPerformed
-        AboutWindow aw = new AboutWindow(rs);
-        aw.pack();
-        aw.setLocationRelativeTo(null);
-        aw.setVisible(true);
-    }//GEN-LAST:event_jButtonInfoActionPerformed
-
-    /**
-     * @param args the command line arguments
-     */
-    public static void main(String args[]) {
-        /* Set the Nimbus look and feel */
-        //<editor-fold defaultstate="collapsed" desc=" Look and feel setting code (optional) ">
-        /* If Nimbus (introduced in Java SE 6) is not available, stay with the default look and feel.
-         * For details see http://download.oracle.com/javase/tutorial/uiswing/lookandfeel/plaf.html 
-         */
-        try {
-            for (javax.swing.UIManager.LookAndFeelInfo info : javax.swing.UIManager.getInstalledLookAndFeels()) {
-                if ("Nimbus".equals(info.getName())) {
-                    javax.swing.UIManager.setLookAndFeel(info.getClassName());
-                    break;
-                }
-            }
-        } catch (ClassNotFoundException ex) {
-            java.util.logging.Logger.getLogger(SetupWindow.class.getName()).log(java.util.logging.Level.SEVERE, null, ex);
-        } catch (InstantiationException ex) {
-            java.util.logging.Logger.getLogger(SetupWindow.class.getName()).log(java.util.logging.Level.SEVERE, null, ex);
-        } catch (IllegalAccessException ex) {
-            java.util.logging.Logger.getLogger(SetupWindow.class.getName()).log(java.util.logging.Level.SEVERE, null, ex);
-        } catch (javax.swing.UnsupportedLookAndFeelException ex) {
-            java.util.logging.Logger.getLogger(SetupWindow.class.getName()).log(java.util.logging.Level.SEVERE, null, ex);
-        }
-        //</editor-fold>
-
-        /* Create and display the form */
-        java.awt.EventQueue.invokeLater(new Runnable() {
-            public void run() {
-                new SetupWindow().setVisible(true);
-            }
-        });
-    }
-
-    // Variables declaration - do not modify//GEN-BEGIN:variables
-    private javax.swing.JButton jButtonExit;
-    private javax.swing.JButton jButtonInfo;
-    private javax.swing.JLabel jLabelCode;
-    private javax.swing.JLabel jLabelCodeImg;
-    private javax.swing.JLabel jLabelHostname;
-    private javax.swing.JLabel jLabelHostnameImg;
-    private javax.swing.JPanel jPanelLabelContainer;
-    // End of variables declaration//GEN-END:variables
-}
